requirements.txt
Flask==3.0.1
openai==1.9.0
Pillow==10.1.0
python-dotenv==1.0.1
SpeechRecognition==3.10.1
